Allows separate peer testing
It modifies the code to allow hybrid KEMTLS connections with different machines, by using the following command-line options: IPserver (IP without the port), tlspeer (client or server) and handshakes (number of iterations for the client). The port was hardcoded to 4433 instead of a random port to ease firewall configurations (because not all ports are open).