Launcher for servers and clients
Adds support to the execution of multiple hybrid KEMTLS servers (one in a single tcp port) and also for the client. The code now executes all of the algorithms present in the hsAlgorithms map. This map is sorted in the client and the server to match the tcp port. They are all in the main package but for this is not a problem for the execution in different machines (or terminal sessions).